Fix test multiple arrows
This test had an error on its `not-repeated-edges?` predicate.
It was comparing the number of unique edges with the unique edges,
and this will ever be false, an example of the wrong comparison is:
```clojure
(= 1 #{{:vertex :E2, :weight 5}})
```

This commit fixes the error by comparing if the number of unique edges
are the same as the number of edges. It should be the same since there
aren't two edges with the same source and target.